Provide a way to execute scripts after running a command
.envrc should contain a list of the commands and the script it should
execute.
There's a lot of magic going on right now:
the name of the script needs to be available in PATH
the name is used in a regex to determine if the command getting run
has a match. If there's a match, the script gets executed.

Probably better to have an array with as key, the regex to check and
value, the scripts to execute.